requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Google Calendar API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Task scheduling
celery==5.3.4
redis==5.0.1

# Email
# Gmail API is included in google-api-python-client

# AI/LLM
openai==1.3.7

# News APIs
newsapi-python==0.2.6

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 